+++ cliente-service/src/main/java/com/bicirepair/cliente_service/model/Cliente.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idCliente;
+    private Long idCliente;
     
     @Column(unique = true, nullable = false)
     private String rutCliente;
